fix: refactor how plugin schema gets exported
