+++ packages/core/tsconfig.json
@@ -3,5 +3,6 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "build"
-  }
+  },
+  "include": ["./**/*.ts", "./src/plugin.schema.json"]
 }
