+++ packages/credential-eip712/package.json
@@ -3,7 +3,10 @@
   "description": "Veramo plugin for working with EIP712 Verifiable Credentials & Presentations.",
   "version": "3.1.0",
   "main": "build/index.js",
-  "exports": "./build/index.js",
+  "exports": {
+    ".": "./build/index.js",
+    "./build/plugin.schema.json": "./build/plugin.schema.json"
+  },
   "types": "build/index.d.ts",
   "scripts": {
     "build": "tsc",
