+++ packages/credential-eip712/src/index.ts
@@ -1,9 +1,2 @@
-/**
- * @public
- */
- import { createRequire } from "module";
- const require = createRequire(import.meta.url);
- const schema = require("../plugin.schema.json");
-export { schema }
 export { CredentialIssuerEIP712 } from './agent/CredentialEIP712.js'
 export * from './types/ICredentialEIP712.js'
