+++ packages/credential-eip712/src/plugin.schema.json
