+++ packages/credential-ld/src/index.ts
@@ -23,7 +23,3 @@ export * from './suites/Ed25519Signature2018.js'
  *
  * @beta This API may change without a BREAKING CHANGE notice.
  */
-import { createRequire } from "module";
-const require = createRequire(import.meta.url);
-const schema = require("../plugin.schema.json");
-export { schema }
