+++ packages/credential-ld/src/plugin.schema.json
