+++ packages/credential-w3c/package.json
@@ -3,7 +3,10 @@
   "description": "Veramo plugin for working with W3C Verifiable Credentials & Presentations.",
   "version": "3.1.4",
   "main": "build/index.js",
-  "exports": "./build/index.js",
+  "exports": {
+    ".": "./build/index.js",
+    "./build/plugin.schema.json": "./build/plugin.schema.json"
+  },
   "types": "build/index.d.ts",
   "scripts": {
     "build": "tsc",
