+++ packages/credential-w3c/tsconfig.json
@@ -12,5 +12,6 @@
     { "path": "../did-jwt" },
     { "path": "../message-handler" },
     { "path": "../utils" }
-  ]
+  ],
+  "include": ["./**/*.ts", "./src/plugin.schema.json"]
 }
