+++ packages/did-comm/package.json
@@ -3,7 +3,10 @@
   "description": "Veramo messaging plugin implementing DIDComm v2.",
   "version": "3.1.4",
   "main": "build/index.js",
-  "exports": "./build/index.js",
+  "exports": {
+    ".": "./build/index.js",
+    "./plugin.schema.json": "./src/plugin.schema.json"
+  },
   "types": "build/index.d.ts",
   "scripts": {
     "build": "tsc",
