+++ packages/did-comm/src/didcomm.ts
@@ -23,7 +23,7 @@ import {
   verifyJWS,
 } from 'did-jwt'
 import { DIDDocument, parse as parseDidUrl, ServiceEndpoint, VerificationMethod } from 'did-resolver'
-import { schema } from './index.js'
+import schema from "./plugin.schema.json" assert { type: 'json' }
 import { v4 as uuidv4 } from 'uuid'
 import * as u8a from 'uint8arrays'
 import {
