+++ packages/did-discovery/src/index.ts
@@ -12,7 +12,3 @@ export * from './types.js'
  *
  * @beta This API may change without a BREAKING CHANGE notice.
  */
- import { createRequire } from "module";
- const require = createRequire(import.meta.url);
- const schema = require("../plugin.schema.json");
-export { schema }
