+++ packages/did-discovery/src/plugin.schema.json
