+++ packages/message-handler/src/message-handler.ts
@@ -4,10 +4,10 @@ import {
   IAgentContext,
   IMessageHandler,
   IHandleMessageArgs,
-  schema,
   CoreEvents,
   IMessage,
 } from '@veramo/core'
+import schema from '@veramo/core/build/plugin.schema.json' assert { type: 'json' }
 import { Message } from './message.js'
 import { AbstractMessageHandler } from './abstract-message-handler.js'
