+++ packages/selective-disclosure/package.json
@@ -3,7 +3,10 @@
   "description": "Veramo plugin to enable the uPort selective disclosure protocol with W3C compatibility.",
   "version": "3.1.4",
   "main": "build/index.js",
-  "exports": "./build/index.js",
+  "exports": {
+    ".": "./build/index.js",
+    "./plugin.schema.json": "./src/plugin.schema.json"
+  },
   "types": "build/index.d.ts",
   "scripts": {
     "build": "tsc",
