+++ packages/selective-disclosure/src/index.ts
@@ -10,15 +10,3 @@
 export { SdrMessageHandler, MessageTypes } from './message-handler.js'
 export { SelectiveDisclosure } from './action-handler.js'
 export * from './types.js'
-
-import { createRequire } from "module";
-const require = createRequire(import.meta.url);
-
-/**
- * The parameter and return types schemas for the {@link @veramo/selective-disclosure#SelectiveDisclosure} plugin
- * methods.
- *
- * @beta This API may change without a BREAKING CHANGE notice.
- */
-const schema = require('../plugin.schema.json')
-export { schema }
