+++ packages/selective-disclosure/src/plugin.schema.json
