+++ packages/test-react-app/.eslintrc.json
@@ -12,7 +12,7 @@
   },
   "parserOptions": {
     "sourceType": "module",
-    "ecmaVersion": 2020,
+    "ecmaVersion": 2022,
     "ecmaFeatures": {
       "jsx": true,
       "modules": true,
