+++ packages/test-react-app/babel.config.json
@@ -1,13 +1,3 @@
 {
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": "current"
-        }
-      }
-    ],
-    "@babel/preset-typescript"
-  ]
+  "plugins": ["@babel/plugin-syntax-import-assertions"]
 }
