+++ packages/test-react-app/package.json
@@ -33,11 +33,11 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "cross-env NODE_OPTIONS=\"--experimental-vm-modules --experimental-json-modules\" react-scripts start",
     "buildz": "react-scripts build",
     "testz": "react-scripts test",
     "ejectz": "react-scripts eject",
-    "test:browser": "cross-env NODE_OPTIONS=--experimental-vm-modules JEST_PUPPETEER_CONFIG=jest-puppeteer.config.cjs jest -c jest-integration.config.cjs"
+    "test:browser": "cross-env NODE_OPTIONS=\"--experimental-vm-modules --experimental-json-modules\" JEST_PUPPETEER_CONFIG=jest-puppeteer.config.cjs jest -c jest-integration.config.cjs"
   },
   "eslintConfig": {
     "extends": [
@@ -59,6 +59,7 @@
   },
   "devDependencies": {
     "@babel/preset-typescript": "7.17.12",
+    "@babel/plugin-syntax-import-assertions": "^7.18.6",
     "@types/eslint": "8.4.3",
     "@types/expect-puppeteer": "4.4.7",
     "@types/jest": "28.1.1",
@@ -67,6 +68,7 @@
     "@types/puppeteer": "5.4.6",
     "@types/react": "18.0.12",
     "@types/react-dom": "18.0.5",
+    "@craco/craco": "^6.4.5",
     "jest": "28.1.1",
     "jest-environment-puppeteer": "6.0.3",
     "jest-environment-puppeteer-jsdom": "6.0.0",
