+++ packages/tsconfig.settings.json
@@ -3,8 +3,8 @@
     "strict": true,
     "preserveConstEnums": true,
     "sourceMap": true,
-    "target": "es2020",
-    "module": "es2022",
+    "target": "esnext",
+    "module": "esnext",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "downlevelIteration": true,
