+++ packages/utils/tsconfig.json
@@ -7,8 +7,8 @@
     "strict": true,
     "preserveConstEnums": true,
     "sourceMap": true,
-    "target": "es2020",
-    "module": "ES2022",
+    "target": "esnext",
+    "module": "esnext",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "downlevelIteration": true,
